+++ includes/parsing.h
@@ -44,12 +44,15 @@ t_token			*new_token(char *value, t_type type);
 void			append_token(t_token *token_list, char *value, t_type type);
 int				operator_token(t_token *token_list, char *line_pos);
 int				value_token(t_token *token_list, char *line_pos);
+
+/* QUOTE HANDLING */
+int				what_quotes(char const *string);
 char			*trim_quote(char *str);
+char			*super_trimmer(char *line);
 //void			print_ttoken(t_token *token); this was for testing
 int				is_quote(char c);
 
 /* EXPANDING */
-int				what_quotes(char const *string);
 char			*expand_vars(char *line, t_data *data);
 int				is_bash_valid(char c);
 char			*append_value(char *new_line, char *var_value, char *var_name);
