+++ source/parsing/parsing_output.c
@@ -12,19 +12,6 @@
 
 #include "../../includes/parsing.h"
 
-char	*trim_quote(char *str)
-{
-	char	*trimmed;
-
-	if (!str)
-		return (NULL);
-	if (what_quotes(str + 1))
-		trimmed = ft_substr(str, 1, ft_strlen(str) - 2);
-	else
-		trimmed = ft_strdup(str);
-	return (trimmed);
-}
-
 char	**args_to_array(t_arg *arg_list)
 {
 	char	**result;
@@ -44,7 +31,7 @@ char	**args_to_array(t_arg *arg_list)
 	while (first->next)
 	{
 		first = first->next;
-		trimmed = trim_quote(first->value);
+		trimmed = super_trimmer(first->value);
 		result[arg_count] = ft_strdup(trimmed);
 		free(trimmed);
 		arg_count++;
